[Fix] gallery teapots notebook [Remove] cylinder test
